+++ examples/completions/rbgp.bash
@@ -7515,7 +7515,7 @@ _rbgp() {
             return 0
             ;;
         rbgp__subcmd__rib__subcmd__advertised)
-            opts="-a -s -j -h --family --explain --rd --addr --token-file --json --no-color --help"
+            opts="-a -s -j -h --family --explain --rd --labeled --addr --token-file --json --no-color --help"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 3 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0